+++ CHANGELOG.md
@@ -58,6 +58,7 @@ v2.0.0
 - Modified `Object.toQueryString` to accept an options object.
 - Modified `Object.fromQueryString` to accept an options object.
 - Added support for Sets in `isEqual`.
+- Added `Array#median`.
 
 
 v1.4.2